llama_cpp_bypass/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
numpy>=1.24.0
pydantic>=2.0.0
llama-cpp-python
